extract test components and textinput
